All Employees views for actions are now in Views Employee, all this controllers are for Employees have become one EmployeeController and all ViewModels for Employee are in one folder
