+++ SiT Access Control System/Data/SACS.Data/appsettings.json
@@ -1,5 +1,5 @@
 ﻿{
   "ConnectionStrings": {
-    "DefaultConnection": "Server=DESKTOP-ICNCU4O;Database=SACS;Trusted_Connection=True;MultipleActiveResultSets=true;TrustServerCertificate=True"
+    "DefaultConnection": "Server=DESKTOP-OGTNCI8;Database=SACS;Trusted_Connection=True;MultipleActiveResultSets=true;TrustServerCertificate=True"
   }
 }
