+++ SiT Access Control System/Web/SACS.Web.ViewModels/Employee/CreateEmployeeAndSummaryViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace SACS.Web.ViewModels
+﻿namespace SACS.Web.ViewModels.Employee
 {
     using System;
     using System.Collections.Generic;
