+++ SiT Access Control System/Web/SACS.Web.ViewModels/Employee/EditEmployeeViewModel.cs
@@ -1,14 +1,13 @@
-﻿namespace SACS.Web.ViewModels
+﻿namespace SACS.Web.ViewModels.Employee
 {
     using System;
     using System.Collections.Generic;
     using System.Linq;
     using System.Text;
     using System.Threading.Tasks;
-
     using SACS.Data.Models;
 
-    public class EmployeeInformationViewModel
+    public class EditEmployeeViewModel
     {
         public string Id { get; set; }
 
@@ -22,6 +21,8 @@
 
         public string Email { get; set; }
 
-        public Department Department { get; set; }
+        public string DepartmentId { get; set; }
+
+        public IEnumerable<Department> Departments { get; set; }
     }
 }
