+++ SiT Access Control System/Web/SACS.Web/Controllers/EmployeeController.cs
@@ -1,28 +1,23 @@
 ﻿namespace SACS.Web.Controllers
 {
     using System;
-    using System.Collections;
-    using System.Collections.Generic;
-    using System.Linq;
-
     using Microsoft.AspNetCore.Mvc;
     using SACS.Data.Models;
     using SACS.Services.Data;
-    using SACS.Web.ViewModels;
+    using SACS.Web.ViewModels.Employee;
 
-    public class AddEmployeeController : Controller
+    public class EmployeeController : Controller
     {
         private readonly IDepartmentService departmentService;
         private readonly IEmployeeService employeeService;
         private readonly ISummaryService summaryService;
         private readonly IUserManagementService userManagementService;
 
-
-        public AddEmployeeController(
-    IDepartmentService departmentService,
-    IEmployeeService employeeService,
-    ISummaryService summaryService,
-    IUserManagementService userManagementService)
+        public EmployeeController(
+            IDepartmentService departmentService,
+            IEmployeeService employeeService,
+            ISummaryService summaryService,
+            IUserManagementService userManagementService)
         {
             this.departmentService = departmentService;
             this.employeeService = employeeService;
@@ -30,21 +25,23 @@
             this.userManagementService = userManagementService;
         }
 
-        public IActionResult Index()
+        // GET: /Employee/Create
+        public IActionResult Create()
         {
             return this.View(new CreateEmployeeAndSummaryViewModel
             {
                 Departments = this.departmentService.GetAll(),
-                Users = this.userManagementService.GetAllUsers(), // implement this if needed
+                Users = this.userManagementService.GetAllUsers(),
             });
         }
 
+        // POST: /Employee/Create
         [HttpPost]
         public IActionResult Create(CreateEmployeeAndSummaryViewModel input)
         {
-            var user = this.userManagementService.GetUserById(input.UserId); // or however you fetch a user
+            var user = this.userManagementService.GetUserById(input.UserId);
 
-            Employee newEmployee = new Employee
+            var newEmployee = new Employee
             {
                 Id = Guid.NewGuid().ToString(),
                 FirstName = input.FirstName,
@@ -52,13 +49,13 @@
                 Position = input.Position,
                 PhoneNumber = input.PhoneNumber,
                 Email = input.Email,
-                Department = this.departmentService.GetDepartmentById(input.DepartmentId),
                 DepartmentId = input.DepartmentId,
-                User = user,
+                Department = this.departmentService.GetDepartmentById(input.DepartmentId),
                 UserId = user.Id,
+                User = user,
             };
 
-            Summary newSummary = new Summary
+            var newSummary = new Summary
             {
                 Id = Guid.NewGuid().ToString(),
                 CurrentState = input.CurrentState,
@@ -66,13 +63,33 @@
                 TotalHoursWorked = input.TotalHoursWorked,
                 Timesabscent = input.Timesabscent,
                 VacationDays = input.VacationDays,
-                Employee = newEmployee,
                 EmployeeId = newEmployee.Id,
+                Employee = newEmployee,
             };
 
             this.employeeService.Add(newEmployee);
             this.summaryService.CreateSummary(newSummary);
+
             return this.Redirect("/");
+        }
+
+        // GET: /Employee/Details/{id}
+        public IActionResult Details(string id)
+        {
+            var currentEmployee = this.employeeService.FindEmployeeById(id);
+
+            var viewModel = new EmployeeInformationViewModel
+            {
+                Id = currentEmployee.Id,
+                FirstName = currentEmployee.FirstName,
+                LastName = currentEmployee.LastName,
+                PhoneNumber = currentEmployee.PhoneNumber,
+                Position = currentEmployee.Position,
+                Department = this.departmentService.GetDepartmentById(currentEmployee.DepartmentId),
+                Email = currentEmployee.Email,
+            };
+
+            return this.View(viewModel);
         }
     }
 }
