--- SiT Access Control System/Web/SACS.Web/Views/EditEmployee/Index.cshtml
@@ -1,40 +1,0 @@
-﻿@model IEnumerable<SACS.Web.ViewModels.EditEmployeeViewModel>
-
-<div class="text-center">
-    <h1 class="display-4">Edit Employees</h1>
-</div>
-
-@foreach (var employee in Model)
-{
-    <form asp-action="Edit" asp-controller="EditEmployee" method="post">
-        <input type="hidden" name="Id" value="@employee.Id" />
-
-        <div class="form-group">
-            <label for="FirstName">First Name</label>
-            <input type="text" class="form-control" id="FirstName" name="FirstName" value="@employee.FirstName" required />
-        </div>
-
-        <div class="form-group">
-            <label for="LastName">Last Name</label>
-            <input type="text" class="form-control" id="LastName" name="LastName" value="@employee.LastName" required />
-        </div>
-
-        <div class="form-group">
-            <label for="Position">Position</label>
-            <input type="text" class="form-control" id="Position" name="Position" value="@employee.Position" required />
-        </div>
-
-        <div class="form-group">
-            <label for="PhoneNumber">Phone Number</label>
-            <input type="text" class="form-control" id="PhoneNumber" name="PhoneNumber" value="@employee.PhoneNumber" required />
-        </div>
-
-        <div class="form-group">
-            <label for="Email">Email</label>
-            <input type="email" class="form-control" id="Email" name="Email" value="@employee.Email" required />
-        </div>
-
-        <button type="submit" class="btn btn-success">Save</button>
-        <a href="@Url.Action("Index", "EditEmployee")" class="btn btn-secondary">Cancel</a>
-    </form>
-}
