+++ SiT Access Control System/Web/SACS.Web/Views/Home/Index.cshtml
@@ -1,6 +1,6 @@
 ﻿@using SACS.Common
 @using SACS.Services.Data
-@model EmployeeListViewModel
+@model SACS.Web.ViewModels.Employee.EmployeeListViewModel
 @{
     this.ViewData["Title"] = "Home Page";
 }
