+++ SiT Access Control System/Web/SACS.Web/Views/Views Employee/AddEmployeeIndex.cshtml
@@ -1,4 +1,4 @@
-﻿@model CreateEmployeeAndSummaryViewModel
+﻿@model SACS.Web.ViewModels.Employee.CreateEmployeeAndSummaryViewModel
 @*
     For more information on enabling MVC for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 *@
@@ -56,42 +56,5 @@
             <select class="form-control" asp-for="DepartmentId" asp-items="departmentInput"></select>
                 <span asp-validation-for="DepartmentId" class="text-danger"></span>
         </div>
-
-        <h2 style="margin-top:50px">Summary Info:</h2>
-
-        <div class="form-group">
-            <label asp-for="CurrentState"></label>
-            <input class="form-control" asp-for="CurrentState">
-            <span asp-validation-for="CurrentState" class="text-danger"></span>
-        </div>
-
-        <div class="form-group">
-            <label asp-for="TimesLate"></label>
-            <input class="form-control" asp-for="TimesLate">
-            <span asp-validation-for="TimesLate" class="text-danger"></span>
-        </div>
-
-        <div class="form-group">
-            <label asp-for="TotalHoursWorked"></label>
-            <input class="form-control" asp-for="TotalHoursWorked">
-            <span asp-validation-for="TotalHoursWorked" class="text-danger"></span>
-        </div>
-
-        <div class="form-group">
-            <label asp-for="Timesabscent"></label>
-            <input class="form-control" asp-for="Timesabscent">
-            <span asp-validation-for="Timesabscent" class="text-danger"></span>
-        </div>
-
-        <div class="form-group">
-            <label asp-for="VacationDays"></label>
-            <input class="form-control" asp-for="VacationDays">
-            <span asp-validation-for="VacationDays" class="text-danger"></span>
-        </div>
-
-
-        <button style="margin-top: 25px" type="submit" class="btn btn-primary btn-lg">Submit</button>
-
-    </div>
 </form>
