+++ SiT Access Control System/Web/SACS.Web/Views/Views Employee/EmployeeInformationIndex.cshtml
@@ -1,5 +1,6 @@
 ﻿@using SACS.Common;
-@model EmployeeInformationViewModel
+using SACS.Web.ViewModels.Employee;
+@model SACS.Web.ViewModels.Employee.EmployeeInformationViewModel
 @{
 }
 <h1 class="text-center">Employee information</h1>
