+++ SiT Access Control System/Web/SACS.Web/appsettings.json
@@ -1,6 +1,6 @@
 {
   "ConnectionStrings": {
-    "DefaultConnection": "Server=DESKTOP-ICNCU4O;Database=SACS;Trusted_Connection=True;MultipleActiveResultSets=true;TrustServerCertificate=True",
+    "DefaultConnection": "Server=DESKTOP-OGTNCI8;Database=SACS;Trusted_Connection=True;MultipleActiveResultSets=true;TrustServerCertificate=True",
     "ApplicationDbContextConnection": "Server=(localdb)\\mssqllocaldb;Database=SACS.Web;Trusted_Connection=True;MultipleActiveResultSets=true"
   },
   "Logging": {
